+++ lib/coinone.rb
@@ -3,6 +3,7 @@ require "coinone/version"
 module Coinone
   require 'coinone/connection'
   require 'coinone/account'
+  require 'coinone/order'
   require 'coinone/public'
   require 'coinone/error'
 end